Extract Class Implemented.
