+++ src/main/java/amidst/fragment/drawer/GridDrawer.java
@@ -19,6 +19,7 @@ public class GridDrawer extends FragmentDrawer {
 	private final StringBuffer textBuffer = new StringBuffer(128);
 	private final char[] textCache = new char[128];
 	private final Zoom zoom;
+	TextDrawer textDrawer = new TextDrawer();
 
 	public GridDrawer(LayerDeclaration declaration, Zoom zoom) {
 		super(declaration);
@@ -36,10 +37,10 @@ public class GridDrawer extends FragmentDrawer {
 		if (isGrid00(gridX, gridY)) {
 			double invZoom = 1.0 / zoom.getCurrentValue();
 			g2d.scale(invZoom, invZoom);
-			updateText(fragment);
-			drawText(g2d);
+			textDrawer.updateText(fragment);
+			textDrawer.drawText(g2d);
 			// drawThickTextOutline(g2d);
-			drawTextOutline(g2d);
+			textDrawer.drawTextOutline(g2d);
 		}
 	}
 
@@ -83,37 +84,5 @@ public class GridDrawer extends FragmentDrawer {
 	@CalledOnlyBy(AmidstThread.EDT)
 	private boolean isGrid00(int gridX, int gridY) {
 		return gridX == 0 && gridY == 0;
-	}
-
-	@CalledOnlyBy(AmidstThread.EDT)
-	private void updateText(Fragment fragment) {
-		textBuffer.setLength(0);
-		textBuffer.append(fragment.getCorner().getX());
-		textBuffer.append(", ");
-		textBuffer.append(fragment.getCorner().getY());
-		textBuffer.getChars(0, textBuffer.length(), textCache, 0);
-	}
-
-	@CalledOnlyBy(AmidstThread.EDT)
-	private void drawText(Graphics2D g2d) {
-		g2d.drawChars(textCache, 0, textBuffer.length(), 12, 17);
-		g2d.drawChars(textCache, 0, textBuffer.length(), 8, 17);
-		g2d.drawChars(textCache, 0, textBuffer.length(), 10, 19);
-		g2d.drawChars(textCache, 0, textBuffer.length(), 10, 15);
-	}
-
-	// This makes the text outline a bit thicker, but seems unneeded.
-	@SuppressWarnings("unused")
-	private void drawThickTextOutline(Graphics2D g2d) {
-		g2d.drawChars(textCache, 0, textBuffer.length(), 12, 15);
-		g2d.drawChars(textCache, 0, textBuffer.length(), 12, 19);
-		g2d.drawChars(textCache, 0, textBuffer.length(), 8, 15);
-		g2d.drawChars(textCache, 0, textBuffer.length(), 8, 19);
-	}
-
-	@CalledOnlyBy(AmidstThread.EDT)
-	private void drawTextOutline(Graphics2D g2d) {
-		g2d.setColor(Color.white);
-		g2d.drawChars(textCache, 0, textBuffer.length(), 10, 17);
 	}
 }
